requirements/main.txt
bcrypt==3.1.0
flask==0.12.2
gunicorn==19.7.1
PyJWT==1.6.0
pyyaml==3.12
